heavy SQL bug hunting session....

+++ libs/server/database/src/repos/sql/control/cases/updateCase.sql
@@ -16,4 +16,4 @@ SET
     extra = ${extra},
     updated_at = NOW()
 WHERE id = ${id}
-RETURNING id::integer;
+RETURNING id;
